fix(general): claim the default crypto only where the runtime can serve Matter
The crypto package installed a Node.js-style implementation into
Environment.default at import time whenever process.getBuiltinModule("crypto")
offered createECDH, without checking that the module works. On Bun and Deno it
therefore claimed the slot with an implementation that has no aes-128-ccm
cipher, and StandardCrypto's own fallback install found the slot taken.

That is what made MATTER_NODEJS_CRYPTO=0 fail on those runtimes: the variable
means "do not install this package's crypto, use the environment default", and
the default was already the broken one.

Auto-detection now probes the API it detects and leaves the slot to
StandardCrypto where a primitive is missing. Where the runtime offers no Web
Crypto there is nothing better, so the detected API is installed regardless and
no runtime ends up without crypto.

nodeCryptoDefect moves to @matter/general alongside the implementation it
probes; @matter/nodejs consumes it.
